llama-index-maxia/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llama-index-maxia"
version = "0.1.0"
description = "LlamaIndex AgentMesh integration for MAXIA — cryptographic trust + on-chain settlement for multi-agent systems"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "MAXIA", email = "contact@maxiaworld.app" },
]
keywords = [
    "llama-index",
    "llamaindex",
    "agentmesh",
    "agent-mesh",
    "maxia",
    "ai-marketplace",
    "multi-agent",
    "ed25519",
    "did",
    "trust",
    "a2a",
    "solana",
    "usdc",
    "crypto",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27",
    "pynacl>=1.5",
    "base58>=2.1",
]

[project.optional-dependencies]
llama_index = [
    "llama-index-core>=0.11",
    "llama-index-agent-agentmesh>=0.2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[project.urls]
Homepage = "https://maxiaworld.app"
Repository = "https://github.com/majorelalexis-stack/maxia"
Documentation = "https://maxiaworld.app/docs"
"API Reference" = "https://maxiaworld.app/api/public/docs"
"MCP Server" = "https://maxiaworld.app/mcp/manifest"
Issues = "https://github.com/majorelalexis-stack/maxia/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/llama_index_maxia"]
